moved db class to module __init__.py
