+++ opendemic/__init__.py
@@ -10,7 +10,7 @@ from flask import Flask, Response
 from flask_cors import CORS
 from apscheduler.schedulers.background import BackgroundScheduler
 from opendemic.webhook.telegram.api_helpers import register_webhook_url, get_telegram_menu, get_telegram_bot_instance
-from opendemic.database.sql_db import RDBManager
+from opendemic.database import RDBManager
 from opendemic.human.human import Human
 import prometheus_client
